refactor(interactors): split interactors modules
