+++ documentation/source/api/use_case_interactors.rst
@@ -1,13 +1,107 @@
 Use case interactors
 ====================
 
-.. inheritance-diagram:: use_case_interactors
+.. inheritance-diagram:: use_case_interactors.base use_case_interactors.create_card use_case_interactors.create_tag use_case_interactors.delete_card use_case_interactors.delete_tag use_case_interactors.list_cards use_case_interactors.read_card use_case_interactors.read_tag use_case_interactors.update_card use_case_interactors.update_tag
+   :top-classes: use_case_interactors.base.UseCaseInteractor
    :parts: 1
    :caption: Relationship diagram
 
+
+Base
+----
+
+.. automodule:: use_case_interactors.base
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
 ---------
 
-.. automodule:: use_case_interactors
+Card
+----
+
+Create
+~~~~~~
+
+.. automodule:: use_case_interactors.create_card
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+Delete
+~~~~~~
+
+.. automodule:: use_case_interactors.delete_card
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+List
+~~~~
+
+.. automodule:: use_case_interactors.list_cards
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+Read
+~~~~
+
+.. automodule:: use_case_interactors.read_card
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+Update
+~~~~~~
+
+.. automodule:: use_case_interactors.update_card
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+---------
+
+
+Tag
+----
+
+Create
+~~~~~~
+
+.. automodule:: use_case_interactors.create_tag
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+Delete
+~~~~~~
+
+.. automodule:: use_case_interactors.delete_tag
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+Read
+~~~~
+
+.. automodule:: use_case_interactors.read_tag
+   :members:
+   :show-inheritance:
+   :member-order: bysource
+
+
+Update
+~~~~~~
+
+.. automodule:: use_case_interactors.update_tag
    :members:
    :show-inheritance:
    :member-order: bysource
